refactor: 전역 상태 파일 분리, 가져오기 업데이트
