+++ src/components/room/ConfirmedButton.tsx
@@ -1,14 +1,14 @@
 'use client';
 import { useGetCalendar } from '@/hooks/useGetCalendar';
-import { useHalfwayDataStore } from '@/store/placeStore';
 import { mostSchedule } from '@/utils/mostSchedule';
 import { useParams } from 'next/navigation';
-import { Button, Modal, ModalBody, ModalContent, ModalFooter, ModalHeader, useDisclosure } from '@nextui-org/react';
-import styles from './ConfirmedButton.module.css';
 import { FormEvent, useState } from 'react';
 import { useQueryUser } from '@/hooks/useQueryUser';
 import { useQuery } from '@tanstack/react-query';
 import { getRoomIsConfirmed, updateRoomData } from '@/api/room';
+import { useHalfwayDataStore } from '@/store/halfwayStore';
+import { Button, Modal, ModalBody, ModalContent, ModalFooter, ModalHeader, useDisclosure } from '@nextui-org/react';
+import styles from './ConfirmedButton.module.css';
 
 const ConfirmedButton = () => {
   const { id: roomId }: { id: string } = useParams();
